requirements.txt
streamlit
requests
python-dotenv
spacy
en-core-web-sm
